Cache stuff, custom headers
This is a major change in the life of this module: caching in browser speeds up a lot the navigation on the site. 304 responses come on the client after the resource is cached.
Also, added the posibility to send custom headers using the last parameter of sendRes method.

Have fun!